pyproject.toml
[project]
name = "py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py-self>=2.1.0",
]
